+++ src/custom_functions/cffi.h
@@ -2,9 +2,9 @@ void cffi_window_edges(
     const int n, //square window size
     long long* edge_idx,
     const int num_elem, //square window size
-    const int* x,
-    const int* y,
-    const int* b,
+    const long long* x,
+    const long long* y,
+    const long long* b,
     bool self_loop,
     long long* edges1,
     long long* edges2
